Add custom build script
The script builds the LaTeX file with 'pdflatex' in a tmp directory, then moves it into place. This is done to prevent PDF viewers from trying to refresh while pdflatex is still compiling, resulting in a black window or even a crashing PDF viewer (in the case of MacOS's preview).